remove root package that contained just one function
