--- go/hashing/hashing.go
@@ -1,36 +1,0 @@
-package hashing
-
-import (
-	"fmt"
-	"sync"
-
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/ethereum/go-ethereum/rlp"
-	"golang.org/x/crypto/sha3"
-)
-
-var hasherPool = sync.Pool{
-	New: func() interface{} { return sha3.NewLegacyKeccak256() },
-}
-
-// RLPHash encodes value, hashes the encoded bytes and returns the hash.
-func RLPHash(value interface{}) (common.Hash, error) {
-	var hash common.Hash
-
-	sha := hasherPool.Get().(crypto.KeccakState)
-	defer hasherPool.Put(sha)
-	sha.Reset()
-
-	err := rlp.Encode(sha, value)
-	if err != nil {
-		return hash, fmt.Errorf("unable to encode Value. %w", err)
-	}
-
-	_, err = sha.Read(hash[:])
-	if err != nil {
-		return hash, fmt.Errorf("unable to read encoded value. %w", err)
-	}
-
-	return hash, nil
-}
